fix(frontend-backend): fixes the interface and the back of the price history

+++ Quinto_Trimestre/TeddyShop - FrontEnd/src/pages/HistorialPrecio/HistorialPrecio.jsx
@@ -24,6 +24,7 @@ import {
   FormControlLabel,
   Checkbox,
 } from '@mui/material';
+import Swal from 'sweetalert2';
 import { Edit, Delete, Info } from '@mui/icons-material';
 import '../PagesStyle.css';
 import { getApiUrl } from '../../utils/apiConfig'
@@ -41,8 +42,7 @@ const HistorialPrecios = () => {
     precio: '',
     fechaInicio: '',
     fechaFin: '',
-    estadoPrecio: true,
-    producto: '',
+    estadoPrecio: true
   });
   const [editingId, setEditingId] = useState(null);
   const [openDetailsDialog, setOpenDetailsDialog] = useState(false);
@@ -58,9 +58,14 @@ const HistorialPrecios = () => {
       const response = await fetch(`${apiUrl}/historialPrecio`, {
         method: 'GET',
         headers: {
-          'Authorization': `Bearer ${token}`, 
+          'Authorization': `Bearer ${token}`,
         },
       });
+  
+      if (!response.ok) {
+        throw new Error(`Error ${response.status}: ${response.statusText}`);
+      }
+  
       const data = await response.json();
       setHistorialPrecios(data);
     } catch (error) {
@@ -89,7 +94,8 @@ const HistorialPrecios = () => {
   
       if (response.ok) {
         fetchHistorialPrecios(); // Actualiza la lista después de crear
-        setNuevoHistorial({ precio: '', fechaInicio: '', fechaFin: '', estadoPrecio: true, producto: '' });
+        resetForm();
+        setNuevoHistorial({ precio: '', fechaInicio: '', fechaFin: '', estadoPrecio: true });
       } else {
         console.error('Error al crear el historial de precio');
       }
@@ -102,22 +108,31 @@ const HistorialPrecios = () => {
   const actualizarHistorialPrecio = async (e) => {
     e.preventDefault();
     try {
-	const token = getAuthToken();
+      const token = getAuthToken();
+      const data = {
+        precio: parseFloat(nuevoHistorial.precio),// Convertir a número
+        fechaInicio: nuevoHistorial.fechaInicio,
+        fechaFin: nuevoHistorial.fechaFin,
+        estadoPrecio: nuevoHistorial.estadoPrecio
+      };
+  
       const response = await fetch(`${apiUrl}/historialPrecio/${editingId}`, {
         method: 'PUT',
         headers: {
           'Content-Type': 'application/json',
-          'Authorization': `Bearer ${token}`, 
+          'Authorization': `Bearer ${token}`,
         },
-        body: JSON.stringify(nuevoHistorial),
+        body: JSON.stringify(data),
       });
   
       if (response.ok) {
-        fetchHistorialPrecios(); // Actualiza la lista después de actualizar
-        setNuevoHistorial({ precio: '', fechaInicio: '', fechaFin: '', estadoPrecio: true, producto: '' });
-        setEditingId(null); // Resetea el ID de edición
+        fetchHistorialPrecios();
+        resetForm();
+        setNuevoHistorial({ precio: '', fechaInicio: '', fechaFin: '', estadoPrecio: true });
+        setEditingId(null);
       } else {
-        console.error('Error al actualizar el historial de precio');
+        const errorText = await response.text();
+        console.error('Error al actualizar el historial de precio:', errorText);
       }
     } catch (error) {
       console.error('Error al enviar los datos:', error);
@@ -126,25 +141,37 @@ const HistorialPrecios = () => {
   
   //Eliminar el historial de Precio
   const eliminarHistorialPrecio = async (id) => {
-    try {
-	const token = getAuthToken();
-      const response = await fetch(`${apiUrl}/historialPrecio/${id}`, {
-        method: 'DELETE',
-        headers: {
-          'Authorization': `Bearer ${token}`, 
-        },
-      });
-  
-      if (response.ok) {
-        fetchHistorialPrecios(); // Actualiza la lista después de eliminar
-      } else {
-        console.error('Error al eliminar el historial de precio');
+    const confirmDelete = await Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'No podrás revertir esta acción',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Sí, eliminar',
+      cancelButtonText: 'Cancelar',
+    });
+
+    if (confirmDelete.isConfirmed) {
+      try {
+        const token = getAuthToken();
+        const response = await fetch(`${apiUrl}/historialPrecio/${id}`, {
+          method: 'DELETE',
+          headers: { 'Authorization': `Bearer ${token}` },
+        });
+
+        if (response.ok) {
+          Swal.fire('¡Eliminado!', 'El historial ha sido eliminado.', 'success');
+          fetchHistorialPrecios();
+        } else {
+          Swal.fire('Error', 'No se pudo eliminar el historial.', 'error');
+        }
+      } catch (error) {
+        Swal.fire('Error', 'Hubo un problema al eliminar el historial.', 'error');
       }
-    } catch (error) {
-      console.error('Error al eliminar:', error);
     }
   };
-  
+
   // Maneja el cambio en el formulario
   const handleChange = (e) => {
     const { name, value } = e.target;
@@ -155,6 +182,12 @@ const HistorialPrecios = () => {
   const iniciarEdicion = (historial) => {
     setNuevoHistorial(historial);
     setEditingId(historial._id);
+  };
+
+  const resetForm = () => {
+   
+    setNuevoHistorial({ precio: '', fechaInicio: '', fechaFin: '', estadoPrecio: true });
+    setEditingId(null);
   };
 
   // Paginación de la tabla
@@ -262,25 +295,31 @@ const HistorialPrecios = () => {
                 '& .MuiTypography-root': { fontSize: '1.2rem' },
               }}
             />
-            <TextField
-              type="text"
-              name="producto"
-              label="ID del Producto"
-              value={nuevoHistorial.producto}
-              onChange={handleChange}
-              fullWidth
-              margin="normal"
-              required
-              variant="outlined"
-              sx={{
-                '& .MuiInputLabel-root': { fontSize: '1.2rem' },
-                '& .MuiInputBase-input': { fontSize: '1.2rem' },
-              }}
-            />
-            <Button type="submit" variant="contained" sx={{ marginTop: 2, fontSize: '1.2rem' }}>
-              {editingId ? 'Actualizar' : 'Crear'}
-            </Button>
-          </form>
+            
+                     <Box display="flex" justifyContent="space-between" mt={2}>
+                       <Button
+                         type="submit"
+                         variant="contained"
+                         sx={{ fontSize: '1.2rem', width: '48%' }}
+                       >
+                         {editingId ? 'Actualizar' : 'Crear'}
+                       </Button>
+                       <Button
+                         type="button"
+                         variant="outlined"
+                         onClick={resetForm}
+                         sx={{
+                           fontSize: '1.2rem',
+                           width: '48%',
+                           backgroundColor: 'transparent',
+                         }}
+                       >
+                         Cancelar
+                       </Button>
+                     </Box>
+                   </form>
+                       
+                   
   
           <Box mt={4}>
             <h2>Lista de Historial de Precios</h2>
@@ -292,26 +331,32 @@ const HistorialPrecios = () => {
                     <TableCell>Fecha Inicio</TableCell>
                     <TableCell>Fecha Fin</TableCell>
                     <TableCell>Estado</TableCell>
-                    <TableCell>Producto</TableCell>
-                    <TableCell>Acciones</TableCell>
+                    <TableCell >Acciones</TableCell>
                   </TableRow>
                 </TableHead>
                 <TableBody>
                   {historialPrecios.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((historial) => (
                     <TableRow key={historial._id}>
-                      <TableCell>{historial.precio}</TableCell>
+                      <TableCell>{new Intl.NumberFormat('es-CO', { style: 'currency', currency: 'COP' }).format(historial.precio)}</TableCell>
                       <TableCell>{new Date(historial.fechaInicio).toLocaleDateString()}</TableCell>
                       <TableCell>{new Date(historial.fechaFin).toLocaleDateString()}</TableCell>
                       <TableCell>{historial.estadoPrecio ? 'Activo' : 'Inactivo'}</TableCell>
-                      <TableCell>{historial.producto}</TableCell>
+                     
                       <TableCell>
-                        <IconButton onClick={() => iniciarEdicion(historial)}>
+                        
+                        <IconButton
+                        color="primary"
+                        onClick={() => iniciarEdicion(historial)}>
                           <Edit />
                         </IconButton>
-                        <IconButton onClick={() => eliminarHistorialPrecio(historial._id)}>
+                        <IconButton 
+                      sx={{ color: "#d33" }}
+                        onClick={() => eliminarHistorialPrecio(historial._id)}>
                           <Delete />
                         </IconButton>
-                        <IconButton onClick={() => handleOpenDetails(historial)}>
+                        <IconButton
+                         color="info"
+                        onClick={() => handleOpenDetails(historial)}>
                           <Info />
                         </IconButton>
                       </TableCell>
@@ -339,7 +384,7 @@ const HistorialPrecios = () => {
           <DialogTitle>Detalles del Historial de Precio</DialogTitle>
           <DialogContent>
             <DialogContentText>
-              <strong>Precio:</strong> {selectedHistorial.precio}
+            <strong>Precio:</strong> {new Intl.NumberFormat('es-CO', { style: 'currency', currency: 'COP' }).format(selectedHistorial.precio)}
               <br />
               <strong>Fecha Inicio:</strong> {new Date(selectedHistorial.fechaInicio).toLocaleDateString()}
               <br />
@@ -347,7 +392,6 @@ const HistorialPrecios = () => {
               <br />
               <strong>Estado:</strong> {selectedHistorial.estadoPrecio ? 'Activo' : 'Inactivo'}
               <br />
-              <strong>Producto:</strong> {selectedHistorial.producto}
             </DialogContentText>
           </DialogContent>
           <DialogActions>
